Se actualizaron los campos del modulo de semen
git-svn-id: svn://pulsatrix.inbio.ac.cr/atta2/trunk@1457 3aa2e761-397d-4522-80f4-092673de1051

+++ ara-ejb/src/java/org/inbio/ara/eao/germplasm/SemenGatheringEAOLocal.java
@@ -40,7 +40,7 @@ public interface SemenGatheringEAOLocal extends BaseLocalEAO<SemenGathering,Long
 
     public List<Long> findBySemenGatheringTime(String time, Long sementalId);
 
-    public List<Long> findByVolume(Long volume, Long sementalId);
+    public List<Long> findByVolume(Double volume, Long sementalId);
 
     public List<Long> findByMotility(Long motility, Long sementalId);
